+++ lizard_fewsnorm/migrations/0013_auto__del_field_event_series__add_field_event_id__add_field_event_comm.py
@@ -1,0 +1,231 @@
+# encoding: utf-8
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding field 'FewsNormSource.data_set'
+        db.add_column('lizard_fewsnorm_fewsnormsource', 'data_set', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['lizard_security.DataSet'], null=True, blank=True), keep_default=False)
+
+    def backwards(self, orm):
+
+        # Deleting field 'FewsNormSource.data_set'
+        db.delete_column('lizard_fewsnorm_fewsnormsource', 'data_set_id')
+
+    models = {
+        'auth.group': {
+            'Meta': {'object_name': 'Group'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '80'}),
+            'permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'})
+        },
+        'auth.permission': {
+            'Meta': {'ordering': "('content_type__app_label', 'content_type__model', 'codename')", 'unique_together': "(('content_type', 'codename'),)", 'object_name': 'Permission'},
+            'codename': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['contenttypes.ContentType']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        'auth.user': {
+            'Meta': {'object_name': 'User'},
+            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'email': ('django.db.models.fields.EmailField', [], {'max_length': '75', 'blank': 'True'}),
+            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Group']", 'symmetrical': 'False', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
+            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'}),
+            'username': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '30'})
+        },
+        'contenttypes.contenttype': {
+            'Meta': {'ordering': "('name',)", 'unique_together': "(('app_label', 'model'),)", 'object_name': 'ContentType', 'db_table': "'django_content_type'"},
+            'app_label': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
+        },
+        'lizard_fewsnorm.event': {
+            'Meta': {'object_name': 'Event', 'managed': 'False'},
+            'comment': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'flag': ('django.db.models.fields.IntegerField', [], {'db_column': "'flags'"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'timestamp': ('django.db.models.fields.DateTimeField', [], {'db_column': "'datetime'"}),
+            'value': ('django.db.models.fields.FloatField', [], {'db_column': "'scalarvalue'"})
+        },
+        'lizard_fewsnorm.fewsnormsource': {
+            'Meta': {'object_name': 'FewsNormSource'},
+            'active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'data_set': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_security.DataSet']", 'null': 'True', 'blank': 'True'}),
+            'database_name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '40'}),
+            'database_schema_name': ('django.db.models.fields.CharField', [], {'max_length': '80', 'null': 'True', 'blank': 'True'}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '50', 'db_index': 'True'})
+        },
+        'lizard_fewsnorm.geolocationcache': {
+            'Meta': {'ordering': "('ident', 'name')", 'object_name': 'GeoLocationCache', '_ormbases': ['lizard_geo.GeoObject']},
+            'active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'data_set': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_security.DataSet']", 'null': 'True', 'blank': 'True'}),
+            'fews_norm_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.FewsNormSource']"}),
+            'geoobject_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['lizard_geo.GeoObject']", 'unique': 'True', 'primary_key': 'True'}),
+            'icon': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'module': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['lizard_fewsnorm.ModuleCache']", 'null': 'True', 'through': "orm['lizard_fewsnorm.TimeSeriesCache']", 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'parameter': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['lizard_fewsnorm.ParameterCache']", 'null': 'True', 'through': "orm['lizard_fewsnorm.TimeSeriesCache']", 'blank': 'True'}),
+            'shortname': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'timestep': ('django.db.models.fields.related.ManyToManyField', [], {'symmetrical': 'False', 'to': "orm['lizard_fewsnorm.TimeStepCache']", 'null': 'True', 'through': "orm['lizard_fewsnorm.TimeSeriesCache']", 'blank': 'True'}),
+            'tooltip': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.location': {
+            'Meta': {'object_name': 'Location', 'managed': 'False'},
+            'area': ('django.db.models.fields.FloatField', [], {}),
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'icon': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'locationkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'locationkey'"}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'parentlocationid': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'related_location': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'shortname': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'tooltip': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'visibilityendtime': ('django.db.models.fields.DateTimeField', [], {}),
+            'visibilitystarttime': ('django.db.models.fields.DateTimeField', [], {}),
+            'x': ('django.db.models.fields.FloatField', [], {}),
+            'y': ('django.db.models.fields.FloatField', [], {}),
+            'z': ('django.db.models.fields.FloatField', [], {})
+        },
+        'lizard_fewsnorm.modulecache': {
+            'Meta': {'ordering': "('ident',)", 'object_name': 'ModuleCache'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'ident': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.moduleinstances': {
+            'Meta': {'object_name': 'ModuleInstances', 'managed': 'False'},
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'moduleinstancekey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'moduleinstancekey'"}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.parameter': {
+            'Meta': {'object_name': 'Parameter', 'managed': 'False'},
+            'attributea': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'attributeb': ('django.db.models.fields.FloatField', [], {}),
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'groupkey': ('django.db.models.fields.CharField', [], {'max_length': '64', 'db_column': "'groupkey'"}),
+            'id': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'parameterkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'parameterkey'"}),
+            'shortname': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'valueresolution': ('django.db.models.fields.FloatField', [], {})
+        },
+        'lizard_fewsnorm.parametercache': {
+            'Meta': {'ordering': "('ident',)", 'object_name': 'ParameterCache'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'ident': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'shortname': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'})
+        },
+        'lizard_fewsnorm.parametergroups': {
+            'Meta': {'object_name': 'ParameterGroups', 'managed': 'False'},
+            'description': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'displayunit': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'groupkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'groupkey'"}),
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'parametertype': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'unit': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.qualifiersetcache': {
+            'Meta': {'ordering': "('ident',)", 'object_name': 'QualifierSetCache'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'ident': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.qualifiersets': {
+            'Meta': {'object_name': 'QualifierSets', 'managed': 'False'},
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'qualifierkey1': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'qualifierkey2': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'qualifierkey3': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'qualifierkey4': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'qualifiersetkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'qualifiersetkey'"})
+        },
+        'lizard_fewsnorm.series': {
+            'Meta': {'object_name': 'Series', 'managed': 'False'},
+            'aggregationperiod': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'location': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'moduleinstance': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'parameter': ('django.db.models.fields.CharField', [], {'max_length': '64'}),
+            'qualifierset': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'serieskey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'serieskey'"}),
+            'timestep': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'unit': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.timeseriescache': {
+            'Meta': {'object_name': 'TimeSeriesCache'},
+            'active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'geolocationcache': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.GeoLocationCache']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'modulecache': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.ModuleCache']"}),
+            'parametercache': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.ParameterCache']"}),
+            'qualifiersetcache': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.QualifierSetCache']", 'null': 'True', 'blank': 'True'}),
+            'timestepcache': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.TimeStepCache']"})
+        },
+        'lizard_fewsnorm.timestepcache': {
+            'Meta': {'ordering': "('ident',)", 'object_name': 'TimeStepCache'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'ident': ('django.db.models.fields.CharField', [], {'max_length': '64'})
+        },
+        'lizard_fewsnorm.timesteps': {
+            'Meta': {'object_name': 'Timesteps', 'managed': 'False'},
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'timestepkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'timestepkey'"})
+        },
+        'lizard_fewsnorm.trackrecordcache': {
+            'Meta': {'ordering': "['datetime']", 'object_name': 'TrackRecordCache', '_ormbases': ['lizard_geo.GeoObject']},
+            'data_set': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_security.DataSet']", 'null': 'True', 'blank': 'True'}),
+            'datetime': ('django.db.models.fields.DateTimeField', [], {}),
+            'fews_norm_source': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.FewsNormSource']"}),
+            'geoobject_ptr': ('django.db.models.fields.related.OneToOneField', [], {'to': "orm['lizard_geo.GeoObject']", 'unique': 'True', 'primary_key': 'True'}),
+            'location': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.GeoLocationCache']", 'null': 'True', 'blank': 'True'}),
+            'module': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.ModuleCache']", 'null': 'True', 'blank': 'True'}),
+            'parameter': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_fewsnorm.ParameterCache']", 'null': 'True', 'blank': 'True'}),
+            'value': ('django.db.models.fields.FloatField', [], {})
+        },
+        'lizard_fewsnorm.users': {
+            'Meta': {'object_name': 'Users', 'managed': 'False'},
+            'id': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '64'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '64', 'null': 'True', 'blank': 'True'}),
+            'userkey': ('django.db.models.fields.IntegerField', [], {'primary_key': 'True', 'db_column': "'userkey'"})
+        },
+        'lizard_geo.geoobject': {
+            'Meta': {'object_name': 'GeoObject'},
+            'geo_object_group': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['lizard_geo.GeoObjectGroup']"}),
+            'geometry': ('django.contrib.gis.db.models.fields.GeometryField', [], {}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'ident': ('django.db.models.fields.CharField', [], {'max_length': '80'})
+        },
+        'lizard_geo.geoobjectgroup': {
+            'Meta': {'object_name': 'GeoObjectGroup'},
+            'created_by': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']"}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'last_modified': ('django.db.models.fields.DateTimeField', [], {'auto_now': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '50', 'db_index': 'True'}),
+            'source_log': ('django.db.models.fields.TextField', [], {'null': 'True', 'blank': 'True'})
+        },
+        'lizard_security.dataset': {
+            'Meta': {'ordering': "['name']", 'object_name': 'DataSet'},
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '80', 'blank': 'True'})
+        }
+    }
+
+    complete_apps = ['lizard_fewsnorm']